pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops"
version = "0.0.1"
description = "MLOps project for Sketch to Image"
authors = [
  { name = "Manh Nguyen", email = "mashwoodev@gmail.com" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9.0"

# DEPENDENCIES
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "albumentations>=1.3.0",
    "python-dotenv>=1.0.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "mlflow>=2.8.0",
    "dvc[s3]>=3.0.0",
    "wandb>=0.15.0",
    "loguru>=0.7.0",
    "typer[all]>=0.9.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "ipython>=8.15.0",
    "jupyter>=1.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
]

# SETUPTOOLS CONFIG
[tool.setuptools]
packages = ["mlops"]

[tool.setuptools.package-data]
mlops = ["config/*.yaml", "**/*.yaml"]

# PYTEST CONFIGURATION
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",                          # Verbose output
    "--strict-markers",            # Require all markers to be registered
    "--cov=mlops",                 # Coverage for mlops package
    "--cov-report=html",           # HTML coverage report
    "--cov-report=term-missing",   # Show missing lines in terminal
    "--cov-report=xml",            # XML report for CI/CD
    "--tb=short",                  # Short traceback format
    "-ra",                         # Show summary of all test outcomes
    "--color=yes",                 # Colored output
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Tests that take a long time to run",
    "skip_ci: Skip in CI environment",
    "gpu: Tests that require GPU",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
# Minimum coverage percentage required
# fail_under = 80

# COVERAGE CONFIGURATION
[tool.coverage.run]
source = ["mlops"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
    "*/venv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# RUFF CONFIGURATION (existing + enhanced)
[tool.ruff]
line-length = 99
src = ["mlops", "tests"]
include = ["pyproject.toml", "mlops/**/*.py", "tests/**/*.py"]
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "htmlcov",
]

[tool.ruff.lint]
extend-select = [
    "I",    # isort
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "C90",  # mccabe complexity
    "C4"
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function call in argument defaults
    "C901",   # Function is too complex
]

[tool.ruff.lint.isort]
known-first-party = ["mlops"]
force-sort-within-sections = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MYPY CONFIGURATION (Type Checking)
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
